refactor(p1): extract ChatInputComposers.kt — TextInputRow + slash panel + @mention panel
Third ChatInput god-class extraction (~720 lines including 4 constants).

What moves to ChatInputComposers.kt (657 lines new):
- internal @Composable TextInputRow — the central composer row (call site:
  main ChatInput composable). Houses TextField, focus state, fullscreen toggle,
  paste handler, and both AnimatedVisibility-wrapped panels.
- private @Composable SlashCommandPanel / SlashCommandRow / SlashCommandEmptyRow
  / QuickMessageButton / SlashCommandLeadingMark — slash-command UI
- private @Composable MentionPanel / MentionRow — @role picker UI
- private sealed interface SlashCommandAction (4 variants: ClearInput,
  CompactContext, OpenUsage, InsertText)
- private data class SlashCommandItem / MentionContext
- private helpers: String.slashCommandQuery, buildSlashCommandItems,
  filterSlashCommandItems, SkillMetadata.toSlashCommandPrompt,
  String.escapeForPromptString, detectMentionContextFor,
  ChatInputState.replaceMention, QuickMessage.slashTitle
- private const MAX_SLASH_COMMANDS, MAX_SLASH_COMMAND_TITLE_CHARS,
  DYNAMIC_SLASH_COMMAND_MIN_QUERY_CHARS, MAX_MENTIONS

Cross-file deps (same package, already `internal`):
- FullScreenEditor (ChatInputAttachments.kt) — TextField fullscreen icon target
- MentionRoleItem / buildMentionRoleItems / filterMentionRoleItems
  (MentionRoles.kt) — @role data + filter

Compose AnimatedVisibility preserved via FQN
(`androidx.compose.animation.AnimatedVisibility(...)` + fadeIn/Out +
slideInVertically/Out + core.tween) — matches the pre-refactor source byte
for byte rather than introducing a top-level animation import to keep this
commit's diff minimal and behavior-identical.

ChatInput.kt: 2551 → 1832 lines (-28%, -720 lines: 716 body + 4 dropped
import-block lines).

Side cleanup — 18 dead imports dropped from ChatInput.kt (verified zero
non-import refs):
  MediaType, ReceiveContentListener, consume, contentReceiver, hasMediaType,
  KeyboardOptions, TextFieldLineLimits, DropdownMenu, HorizontalDivider,
  TextField, TextFieldDefaults, produceState, onFocusChanged, ImeAction,
  FullScreen, SubAgentMode, getQuickMessagesOfAssistant, SkillManager,
  SkillMetadata, QuickMessage, Locale.

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
